'error handling added to delete comment functions

+++ models.js
@@ -14,8 +14,6 @@ exports.fetchArticles = (topic, sort, order) => {
   const acceptedTopics = ['mitch', 'cats', 'paper', '', undefined];
   const acceptedSortBys = ['article_id', 'title', 'topic', 'author', 'body', 'created_at', 'comment_count', 'article_img_url', 'votes', '', undefined];
   const acceptedOrderBys = ['desc', 'asc', '', undefined];
-  console.log(acceptedTopics.includes(topic))
-
 if(!acceptedTopics.includes(topic) || !acceptedOrderBys.includes(order) || !acceptedSortBys.includes(sort)){
   return Promise.reject({status: 400, msg: 'bad query request'})
 }
@@ -143,11 +141,20 @@ exports.fetchUsers = () => {
 
 
   exports.eraseComment = (comment_id) => {
+    const acceptedInput = new RegExp(/^\d+(?:\.\d{1,2})?$/);
+    console.log(acceptedInput.test(+comment_id))
+
+  if (acceptedInput.test(comment_id) === false) {
+    return Promise.reject({ status: 400, msg: "bad delete request" });
+  }
+
     const query = `DELETE FROM comments WHERE comment_id = $1
     RETURNING *;`
     return db.query(query, [comment_id])
     .then((comments) => {
-      console.log(comments.rows[0])
+      if(comments.rowCount === 0){ 
+        return Promise.reject({status: 404, msg: 'comment not found'})
+      }
       return comments.rows[0]
     })
   }
